Make intrgral attributes signed
Cooler does not yet properly handle uint attributes.
See https://github.com/open2c/cooler/pull/323 for more details

+++ src/coolerpp_write_impl.hpp
@@ -294,7 +294,7 @@ inline void File::update_indexes(PixelIt first_pixel, PixelIt last_pixel) {
 
   const auto last_bin_written = this->get_last_bin_written();
 
-  auto nnz = *this->_attrs.nnz;
+  auto nnz = static_cast<std::uint64_t>(*this->_attrs.nnz);
   PixelCoordinates first_pixel_in_row{this->_bins, last_bin_written.chrom.name,
                                       last_bin_written.start, last_bin_written.start};
 
@@ -309,7 +309,7 @@ inline void File::update_indexes(PixelIt first_pixel, PixelIt last_pixel) {
 
 inline void File::write_indexes() {
   assert(this->_attrs.nnz.has_value());
-  this->index().finalize(*this->_attrs.nnz);
+  this->index().finalize(static_cast<std::uint64_t>(*this->_attrs.nnz));
   File::write_indexes(this->dataset("indexes/chrom_offset"), this->dataset("indexes/bin1_offset"),
                       this->index());
 }
